refactor(meeting_facilitator): split handoff.rs into mod+persistence (#1266)
handoff.rs 416 -> handoff/{mod.rs ~240, persistence.rs ~189}.
